+++ R/ipPackage.R
@@ -1,14 +1,14 @@
 #' Tabulate an IP's package downloads.
 #'
 #' From RStudio's CRAN Mirror http://cran-logs.rstudio.com/
-#' @param ip Numeric. ip_id.
+#' @param ip Numeric. ip_id. Positive integer.
 #' @param date Character. Date. "yyyy-mm-dd". NULL uses latest available log.
 #' @param all.filters Logical. Master switch for filters.
 #' @param ip.filter Logical.
 #' @param small.filter Logical. TRUE filters out downloads less than 1000 bytes.
 #' @param sequence.filter Logical.
 #' @param size.filter Logical.
-#' @param sort Logical. Sort by download count.
+#' @param sort.count Logical. Sort by download count.
 #' @param memoization Logical. Use memoization when downloading logs.
 #' @param multi.core Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.
 #' @note ip = 10 is a tw top-level domain on 2020-07-09.
@@ -16,59 +16,52 @@
 
 ipPackage <- function(ip = 10, date = NULL, all.filters = FALSE,
   ip.filter = FALSE, small.filter = FALSE, sequence.filter = FALSE, 
-  size.filter = FALSE, sort = TRUE, memoization = TRUE, multi.core = FALSE) {
+  size.filter = FALSE, sort.count = TRUE, memoization = TRUE, 
+  multi.core = FALSE) {
 
-  ymd <- logDate(date)
-  cran_log <- fetchCranLog(date = ymd, memoization = memoization)
-  cran_log <- cleanLog(cran_log)
+  if (!curl::has_internet()) stop("Check internet connection.", call. = FALSE)
+  if (ip < 1)  stop("Invalid IP. Use a positive integer.")
 
   cores <- multiCore(multi.core)
+  if (.Platform$OS.type == "windows" & cores > 1) cores <- 1L
 
-  # N.B. using pkg_specific_filters not recommended!
+  file.url.date <- logDate(date)
+  cran_log <- fetchCranLog(date = file.url.date, memoization = memoization)
+  cran_log <- cleanLog(cran_log)
+  ymd <- rev_fixDate_2012(file.url.date)
+  
   if (all.filters) {
-    ip.filter <- TRUE
-    small.filter <- TRUE
-    # sequence.filter <- TRUE
-    # size.filter <- TRUE
+    ip.filter = TRUE
+    small.filter = TRUE
+    sequence.filter = TRUE
+    size.filter = TRUE
   }
-
-  pkg_specific_filters <- c(sequence.filter, size.filter)
-
-  if (ip.filter) {
-    cran_log <- ipFilter(cran_log, multi.core = cores)
-  }
-
-  if (any(pkg_specific_filters)) {
-    pkgs <- unique(cran_log$package)
-
-    out <- parallel::mclapply(pkgs, function(p) {
-      cran_log[cran_log$package == p, ]
-    }, mc.cores = cores)
-
-    if (small.filter) {
-      out <- smallFilter(out)
-    }
-
-    if (sequence.filter) {
-      arch.pkg.history <- parallel::mclapply(pkgs, function(x) {
-        tmp <- packageHistory(x)
-        tmp[tmp$Date <= ymd & tmp$Repository == "Archive", ]
-      }, mc.cores = cores)
-
-      out <- parallel::mclapply(seq_along(out), function(i) {
-        sequenceFilter(out[[i]], arch.pkg.history[[i]])
-      }, mc.cores = cores)
-    }
-
-    if (size.filter) out <- sizeFilter(out, pkgs, cores)
-    cran_log <- do.call(rbind, out)
-
+  
+  if (small.filter) cran_log <- smallFilter(cran_log)
+  if (ip.filter) cran_log <- ipFilter(cran_log, multi.core = cores)
+  
+  if (!ip %in% cran_log$ip_id) {
+    stop("IP not found in log.")
   } else {
-    if (small.filter) cran_log <- cran_log[cran_log$size >= 1000L, ]
+    cran_log <- cran_log[cran_log$ip_id == ip, ]
+    pkgs <- unique(cran_log$package)
+    
+    out <- parallel::mclapply(pkgs, function(p) {
+      pkg.data <- cran_log[cran_log$package == p, ]
+      if (sequence.filter) pkg.data <- sequenceFilter(pkg.data, p, ymd)  
+      if (size.filter) pkg.data <- sizeFilter(pkg.data, p)
+      pkg.data$date.time <- dateTime(pkg.data$date, pkg.data$time)  
+      pkg.data <- pkg.data[order(pkg.data$date.time), ]
+      pkg.data$date.time <- NULL
+      pkg.data
+    }, mc.cores = cores)
+    
+    cran_log <- do.call(rbind, out)
   }
-
-  cran_log <- cran_log[cran_log$ip_id == ip, ]
-  freqtab <- table(cran_log$package)
-  if (sort) sort(freqtab, decreasing = TRUE)
-  else freqtab
+  
+ if (nrow(cran_log) > 0) {
+    freqtab <- table(cran_log$package)
+    if (sort.count) sort(freqtab, decreasing = TRUE)
+    else freqtab
+  } else stop("IP not found in log.")
 }